bug 8009: fixed by implementing: http://sarek.l7tech.com/mediawiki/index.php?title=CentraSite_ActiveSOA#Show_select_wsdl:port_dialog
Also updated UDDI searches to batch keys where possible to reduce round trips to registry. Added new system property for batch size for when retrieving batches of tModels. Added new check box in search uddi dialog to allow user to not require the WSDL URL to be displayed in search results. This dramatically increases search time, the URL is then retireved when the selection is made.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@24963 219f1c6e-4a0e-43c8-96f2-6e749be13fd4