bug 7983: The Gateway now installs Crypto-J 4.1.0.1 as the least-preference JCE provider when the Gateway is running in non-FIPS mode.  In FIPS mode, the Gateway continues to install Crypto-J 4.0 FIPS as the highest-preference JCE provider.
This change causes the Gateway to (when not in FIPS mode) use crypto implementations from the JDK, instead of from Crypto-J, for all crypto operations other than elliptic curve cryptography.  In this configuration, performance of RSA and DHE_RSA cipher suites is back to 5.0 levels.

Elliptic curve functionality should be unaffected by this change.  For 5.2, TLS client authentication with an elliptic curve client certificate should continue to be supported.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@25044 219f1c6e-4a0e-43c8-96f2-6e749be13fd4