bug 7210: Fixed by prompting an optional dialog to resolve policy fragment conflicts if there are some imported policies are already imported in the system.
The dialog message is "The imported policy contains an embedded fragment with name 'XYZ' that already exists in the system.  This policy importer will use the already existing policy fragment in place of the embedded fragment."

There are two buttons, OK and Cancel in the dialog.  If the user chooses OK, the importer will continue policy importing.  Otherwise, the process of policy importing will be aborted.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@25239 219f1c6e-4a0e-43c8-96f2-6e749be13fd4