bug 5784: Fixed by adding a new flag to indicate whether the provider type is customized or not.
If the provider type is set as "(CUSTOM)" when a queue is created, no matter whether the Initial Context Factory is set as specific context factory classes (such as "com.sun.jndi.ldap.LdapCtxFactory", etc), when the user edits the queue at next time, the provider type will not automatically change to other types and will keep the same as "(CUSTOM)".  If the user changes the provider type from "(CUSTOM)" to other types and saves this queue, when the user edits the queue at next time, the changed provider type will be whatever the user changed at last time.

We still preserve the documentation mentioned in the bug 4986, "Changing from a specific provider type to “(Custom)” has no effect on the queue properties."

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@25682 219f1c6e-4a0e-43c8-96f2-6e749be13fd4