bug 8586: Fixed by improving the error message when can't load WSDL files from SSL URL because we don't trust the server. The error message format is "Could not fetch the WSDL at location 'https://XYZ': Unable to obtain HTTP response from /ABC: Certificate not verified.
The SSG will log a warning for this error.  The log message is "WARNING: 2034: Unable to build path for Certificate (0} : {1}", where {0} represents the subject DN info and {1} represents the error detail such as either "unable to find valid certification path to requested target" or "the trustAnchors parameter must be non-empty".

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@25779 219f1c6e-4a0e-43c8-96f2-6e749be13fd4