bug 8657: Increased the default WS-SecureConversation security context shared secret size from 16 to 32 bytes; introduced system property com.l7tech.security.secureconversation.defaultSecretLengthInBytes to configure it on the Gateway.
Increased the default derived key length from 16 to 32 bytes and introduced the system property com.l7tech.security.secureconversation.defaultDerivedKeyLengthInBytes to configure it on the Gateway and the XVC.

The WssDecorator now respects the configured symmetric algorithm when encrypting using WS-SecureConversation.  If the key derivation did not produce enough key material for the highest-preference cipher decoration will now fail rather than silently fall back to AES 128.  (Since the SSM GUI always configures the assertion with a cipher preference order from weakest to strongest it should never be the case in practice that we fail on the highest-preference cipher when we could have switched to a weaker cipher.) 

Simplified RequireWssEncryptedElement's handling of algorithm preference order and added a unit test.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@25903 219f1c6e-4a0e-43c8-96f2-6e749be13fd4