Policy tracing:  The Service Properties and Policy Properties dialogs now both show the policy GUID (as a read-only field).
Enabling or disabling the trace bit for a service now requires a role with permission to UPDATE all Published Service entities.  There is a separate admin method for changing the tracing bit.  The regular save() admin methods will refuse to change an existing tracing bit, and will refuse to set the tracing bit for a brand-new service entity.

Introduced new special mutable context variable "trace.out".  This behaves like a regular user-created context variable, in that it can be set to any type, but it is already set before the trace policy starts.  Whatever is left in the special context variable "trace.out" when the trace context is closed is now logged by TracePolicyEvaluator at level FINER.  To have this information saved to a log file on the Gateway, do the following:

1. In Manage Log/Audit Sinks, create a File log sink named "trace" that records severity All for the Gateway Log category.  

2. Edit the cluster property "log.levels" and add a line "com.l7tech.server.trace.TracePolicyEvaluator.level = FINER".

3. Configure your trace policy to accumulate any desired trace information in the context variable "trace.out".  


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@26082 219f1c6e-4a0e-43c8-96f2-6e749be13fd4