Initial implementation of nCipher JceProviderEngine and KeyStore.
Currently requries that the nCipher software and drivers are manually installed, the security world is manually initialized or programmed, the nCipherKM.jar file is manually copied to $JAVA_HOME/jre/lib/ext, the gateway user is added to the nfast group, and the com.l7tech.common.security.jceProviderEngine is manually set.  

Currently uses module protection, with ignorePassphrase=true.  This means any softwar running on the appliance as a user in the nfast group can currently use the keystore without a password.

No rollback of written key blobs is currently supported in the event of rollback of the enclosing Hibernate write transaction in which the keystore mutation takes place.  However, because new (though not changed) key blobs are merged non-destructively (replicated) to/from the database, this is still much cleaner and safer than the current approach with the SCA 6000 keystore; specifically, it should not be possible to end up with missing keys or an unusable keystore no matter how badly the transaction goes sideways.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@26179 219f1c6e-4a0e-43c8-96f2-6e749be13fd4