bug 9260: The default XPath expression for new XpathBasedAssertion instances created in the SSM within a SOAP policy is now "/s:Envelope/s:Body", with the "s:" prefix declared as the SOAP 1.2 namespace if the service is known and its SOAP version is set to SOAP 1.2 and the SOAP 1.1 namespace otherwise.  Assertions created the old way continue to work as expected.
The namespace map of an XPath based assertion no longer has all the namespaces from the WSDL dumped into it as soon as an operation is selected.  Now, the namespace map is only changed when the user elects to replace the current expression with an automatically generated expression by clicking on a node in a sample message.  Only namespace declarations that are actually used by the generated expression are added to the namespace map.

There are no longer "fixed" namespaces which cannot be edited.  The entire namespace map is now under the control of the user.

The "wsp" and "l7p" prefixes are no longer treated specially by the xpath based assertion properties dialog.  These prefixes are no longer added automatically to the namespace map, and if added by hand with arbitrary values, the user-supplied prefixes and URIs will be preserved and work as expected.  (At least part of this was already fixed earlier by changeset 23626.)

The NamespaceMapEditor now displays a "Remove All Unused" button when invoked from the XPath based assertion properties dialog.  When pressed, it cleans out any namespace declarations that are not visibly used by the current XPath expression.

The SOAP message generator now always uses the "s:" prefix for the SOAP envelope namespace URI regardless of SOAP version (though it declares "s:" with the appropriate URI for the SOAP version, of course).

Implemented ${foo.soap.version} and ${foo.soap.envelopens} context variables for all Message variables, including ${request} and ${response}.

The Gateway now checks that the SOAP version of a request is compatible with the target service's configured SOAP version at the end of service resolution, as part of validating the SOAP operation.  This validation is skipped if the service is in lax mode.

Fix: The validator warning for an XPath expression using a mismatching SOAP namespace now only fires if the expression is actually visibly using the mismatching namespace URI (rather than merely retaining an unused declaration for it in its namespace map).

Fix: The NamespaceMapEditor dialog's table no longer allows multi-selection, since none of its actions ever supported it.

Fix: The XPath properties dialog no longer modifies the assertion's namespace map if the dialog is canceled.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@26917 219f1c6e-4a0e-43c8-96f2-6e749be13fd4