bug 8866: Fixed by adding "Maximum Expiry Time" in the "Require SAML Token Profile" assertion to enforce to check if the SAML token is expired.
The valid value of "Maximum Expiry Time" is between 0 and 100 years.  The default value is "0", which means not enforcing to check token expiration.  Any invalid value will disable the Finish or OK button in the "SAML Token Profile Wizard".

If the SAML token is validated as expired, then the policy consumption will fail and the validator will log and audit: "SAML token is expired when constrained to maximum allowed lifetime".  The Audit Code is 6108 with the WARNING Level.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@27483 219f1c6e-4a0e-43c8-96f2-6e749be13fd4