bug 9749: The WSS decorator now supports whole-element (rather than element-content), with signature encryption offered as an easy-to-activate special case.
The "Encrypt Element" assertion GUI has had radio buttons added to control whether to encrypt only the element contents or to encrypt the open and closing tags as well.  The default is to encrypt only the element contents as we did previously.

The "Configure WS-Security Decoration" assertion now has a checkbox for "Encrypt Signature".  This causes the signature to be included in the elements to encrypt, marked as requiring whole-element encryption.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@27662 219f1c6e-4a0e-43c8-96f2-6e749be13fd4