bug 9522: Per the comment #5 in bug 9522, fixed some policy validation warnings for the assertions, Sign Element and Configure WS-Security Configuration on the response decoration part in the default policy of a STS Internal Service.  No need to fix the validation warning for the Encrypt Element assertion.
Set a new flag "Flag_Enable_Response_Decoration" in the default policy to determine if enabling or disabling the RSTR response decoration.  The flag will be turned off if the request credentials rely on any Transport Level Security or turned on if the request credentials rely on Message Level Security.  By default, the flag is turned off.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@27861 219f1c6e-4a0e-43c8-96f2-6e749be13fd4