bug 9939: Outbound HTTP configuration options now allow configuration of the TLS cipher suites to enable.  The HTTP routing assertion now allows configuration of the outbound TLS version and cipher suites to enable.
Extracted the guts of the cipher suite configuration GUI from SsgConnectorPropertiesDialog into a new class CipherSuitGuiUtil.  Added a utility dialog CipherSuiteDialog that can be popped up in the SSM to configure a cipher suite list.   

Note that the "https.protocols" and "https.cipherSuites" system properties are applied during hostname verification, latei in the connection establishment process, so if either of these system properties is set it will override any corresponding setting specified in a routing assertion.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@27903 219f1c6e-4a0e-43c8-96f2-6e749be13fd4