Keystore-protected master passphrase progress.
DefaultKeyImpl now logs in addition to throwing if there is a failure while preparing the default SSL key (since such failures are prone to being overwritten by subsequent failures and never logged).

The MasterPasswordChanger's kmp facility now provides a listKeystoreContents option which behaves like a version of keytool --list with drop-in support for kmp config profiles as well as cleaner output.

The Gateway and the ncipher config script now keep track of the keystore ID in use in the Gateway database.  When configuring the Gateway to use a manually-preprogrammed security world, the ncipher config script will now prompt the user to choose if more than one keystore is present in kmdata/local, and will warn if there is a conflict with the keystore ID in the database.

Enabling Gateway use of an nCipher HSM now converts the node to use a keystore-protected master passphrase as a side-effect.  Disabling nCipher HSM migrates the kmp back to omp and deletes the kmp key blobs from kmdata.  

The Gateway must now have use of nCipher HSM turned off before the ncipher config script will permit changing the security world configuration.

Keystore implementations that use a keystore_file row for backing may now atomically update properties at the same time as (or instead of) updating the databytes column upon keystore mutation.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@27979 219f1c6e-4a0e-43c8-96f2-6e749be13fd4