Secure password expansion now checks that referenced passwords are enabled for use via context variable, when used for template expansion (simulating variable expansion with a PEC that allows only ${secpass} variables).  This is how it would typically be accessed from a server assertion that needed to expand a password variable in its constructor.
It continues to allow references to passwords that don't allow use via context variable as long as this is a password-only reference (allowing only a single reference of the form "${secpass.FOO.plaintext}" with no additional surrounding text).  This is how it would typically be accessed when a (non-assertion) entity password field is being processed.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@28191 219f1c6e-4a0e-43c8-96f2-6e749be13fd4