bug 4458: It is now possible to configure the audit signing key used by the Gateway.  It defaults to the same as the SSL key.  The SSM will use the Gateway's current default audit signing key when verifying signatures in the audit viewer window.  You can change the default audit signing key on the fly on a single-node cluster and both the Gateway and SSM (the one that made the change) will pick up the change immediately.  However, this is not guaranteed to work in a multi-node cluster without restarting the other nodes.
The SSM (and the admin interface) now supports an arbitrary number of SpecialKeyType items instead of having hardcoded support for three items.  The GUI will adjust itself based on what values are available in this enum.  Changing of assigned special key roles by the SSM is now done through TrustedCertAdmin.  However, on the back end, the metadata is still stored in cluster properties for now.

Added a hack to DialogDisplayer to attempt to work-around a dialog size issue when a multiline safe confirmation dialog is displayed as a sheet.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@28243 219f1c6e-4a0e-43c8-96f2-6e749be13fd4