bug 9602:  It is now possible to use non-SOAP sign element to create detached signatures which can later be added to the same or a different document.  If a detached signature is created, it will be a single signature that covers all elements matching the XPath of elements to sign.  It will be left in the specified context variable and will not be added to the document.  It may optionally include the Enveloped transform, in case it is to be (manually) added to the document as a descendant of one of the signed elements.
Non-SOAP sign element now supports creating signature referenes that use the implicit empty string "" reference value to refer to the document element of the document being signed.  To create such a signature, set the XPath to a value (ie, /*) that includes the document element among its matches, and then use a custom ID attribute that is completely empty.

The "Remove Element" assertion has been enhanced and is now the "Add or Remove Element" assertion.  It can be used to add a new element to the target document as either the first child, last child, previous sibling, or next sibling of the targeted existing element.  This can be combined with the new Sign Element behavior in order to more easily add a detached signature to a document.

For convenience, the "Remove Element" assertion now directly recognizes an expression such as "requestXpath.elements" if used as a variable name (though not anything more complicated such as "requestXpath.elements[0]").

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@28346 219f1c6e-4a0e-43c8-96f2-6e749be13fd4