bug 10085: Fixed by displaying correct information in the details pane for a selected audit in the message table in the audit viewer.
For DEV note, the issue described by bug 10085 was caused by ClusterLogWorker in AuditLogTableSorterModel making extra call on realModel to fire a tableDataChanged event.  When the user is selecting an audit in the message table, realModel.fireTableDataChanged() will clear all table selections, this will cause that the selected row index is set to -1, which is equivalent to no selection in the message table at all.  In GUI, the details pane still displays the details of the last selected audit rather than the current really-selected audit's details.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@28398 219f1c6e-4a0e-43c8-96f2-6e749be13fd4