bug 9161: The CancelableOperationDialog for slow remote calls made on the Swing event thread now will no longer appear until Swing event processing has been blocked for a full ten seconds.  This makes it less useful as a means for preserving SSM GUI responsiveness, but it is still useful for providing a "Cancel" button as an escape hatch in exceptional circumstances.  It should also pop up far less often and be generally less annoying.
The CancelableOperationDialog for slow remote calls now centers on the SSM main window and should no longer steal focus from other apps when it appears.

Dialogs that center on their parent window, including the CancelableOperationDialog, will no longer appear against the right edge of the leftmost monitor when the main SSM window is placed on a different monitor on a multi-monitor system.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@28802 219f1c6e-4a0e-43c8-96f2-6e749be13fd4