bug 11251: Added some countermeasures for the "How to Break XML Encryption" attack.
(1) A decryption key blacklist.  If a decryption attempt using given secret key bytes has failed too many times on the current node, we blacklist those key bytes from further attempts to decrypt anything.  This feature can be disabled with a system property.
  
(2) Decryption attempts that fail, for any reason, no longer report failure.  Instead, they succeed but replace the encrypted value with a dummy element.  The hope is that a later signature check will cause the message to be rejected.  This feature can be disabled with a system property.  However, even without this feature enabled, decryption attempts using the wrong key can still happen to result in well-formed XML (though the probability quickly becomes infinitesimal as the ciphertext gets larger than a block or two).  This change also addresses Bug #9946

This approach is in need of further review and may change if necessary.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@29723 219f1c6e-4a0e-43c8-96f2-6e749be13fd4