bug 4772: Productized WSS Digest assertion per func spec at http://sarek/mediawiki/index.php?title=WSS_Digest
This version does not act like a credential source assertion -- it simply requires that a WSS Digest token was present with a matching username and password (which may use variables, including secure password references).  This version can check that a timestamp and/or nonce are present, but does not help you check whether the timestamp has expired, and does not help you enforce that the nonce is not reused.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@30365 219f1c6e-4a0e-43c8-96f2-6e749be13fd4