Adding support for a generic entity manager to manage entities whose concrete classes are not necessarily known at Gateway compile time.  This allows modular assertions to manage their own entities.
Added a demo entity manager dialog to the WhichModuleAssertion that does CRUD on a simple demo entity.

There is currently an issue where the RBAC interceptor fails to enforce a method using an entity OID for the relevantArg (such as DELETE_BY_OID) when the entity is a generic entity.  This can be worked around by sending up the entire entity to delete and using a DELETE_ENTITY stereotype instead for the time being.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@30414 219f1c6e-4a0e-43c8-96f2-6e749be13fd4