bug 11733: MQ Native Policy Export/Import initial implementation: (1) In Policy Exporter, the following reference details are included in the "MqNativeExternalReference" element in the policy xml - OID - ConnectorName - Host - Port - QueueManagerName - ChannelName - QueueName
(2) In Policy Importer, if a MQ Native queue referenced by an external reference is missed, then a Resolve External Dependencies Wizard will display a Wizard Step Panel to allow user to resolve the missed reference by three options:
- Change assertions to use this queue
- Remove assertions that refer to the missing queue
- Import erroneous assertions as-is

In the wizard step panel, there is a button, "Create MQ native Queue", which can create a new queue. In this case, the first action "Change assertions to use this queue" will be enabled and selected.

(3) For DEV note only, this bug fixing implemented a new way for Modular Assertion Policy Export/Import, in which a new ExternalReferenceFactory class is introduced.  When the gateway starts and loads each modular assertion, each corresponding ExternalReferenceFactory will be registered. In the policy exporter, each ExternalReferenceFactory will be asked if it wants to create a reference for modular assertion in the policy. In the policy importer, each ExternalReferenceFactory will be asked to parse reference for external reference element in the policy export document.  If an unknown external reference cannot be resolved, then ExternalReferenceFactory will return a WizardStepPanel to ResolveExternalPolicyReferencesWizard.  Thus, both of modular external reference and wizard step panel can be bundled into the modular assertion package.  This new way may achieve all related resources grouped together in a single modular package.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@30711 219f1c6e-4a0e-43c8-96f2-6e749be13fd4