bug 11984: Attacked policy path building with a profiler and made a number of changes that significantly increase the performance of policy path building (while making it use very slightly more memory per path).  As a result, also increased the default policy path limit from 500,000 to 2,000,000.
AssertionPath now records its size when created, since this can never change.  containsSibling() no longer builds a temporary path array, and now short circuits on success.  Switched pathStack from a Stack to an ArrayDeque.  Introduced Assertion.isSibling().  

When an attempt is made to download a policy via the policy download service that is too complex to analyze quickly, the Gateway will now assume the policy does not allow anonymous access.

There are still a number of areas for possible improvement but this should be a good improvement for now.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@30938 219f1c6e-4a0e-43c8-96f2-6e749be13fd4