bug 11441: Add support for X509SubjectName for cert lookup in the NonSoapVerifyElement assertion if X509Data is present but none of the other less-ambiguous fields are present.
Also fixed a JDK7 incompatibility due to the addition of the SunEC provider in JDK7 -- the SunEC provider is unable to use the Apache cert in one of the sample messages because it does not use a well-known named curve.  The work-around is to remove the SunEC provider for the test, allowing it to go back to using BouncyCastle's EC support as it did in JDK6.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@31226 219f1c6e-4a0e-43c8-96f2-6e749be13fd4