bug 12148: Message targetable variables that permit use with context variables of non-Message type now behave more reasonably for value types other than String.  For the "Store to Cache", "Validate JSON Schema", "Apply JSON Transformation" and "Route via Raw TCP" assertions, the following logic is used:
If target is a CharSequence, use its toString() value.  Else if target is a Collection or Array with exactly one member, recursively check its member.  Else fail with NoSuchVariableException, which will cause the assertion to fail.

For the "Evaluate Regular Expression" assertion, different logic is used for maximum backward compatibility with the pre-Fangtooth behavior when the regex assertion was used with a Collection variable:

If target is a CharSequence or AbstractCollection, use its toString() value.  Else if target is an array or other collection type, convert it to an ArrayList and use its toString() value.  Else use its raw toString() value.

This will result in targets similar to the following:

For a String like "foo", the target will just be "foo".

For a multivalued variable with one element, like {"foo"}, the target will be a string similar to "[foo]".

For a multivalued variable with more than one element, like {"foo", "bar", "baz"}, the target will be a string similar to "[foo, bar, baz]".

For any other object, the target will be the value it returns from toString(), in case it is what is desired (even though it often will not be).

We should consider providing a way to disable this behavior in the future (hopefully without adding more UI noise to the main screen of the regex properties dialog).


Removed a misleading constructor from NoSuchVariableException and fixed an incorrect usage.   Made some minor JavaDoc and @NotNull fixes.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@31294 219f1c6e-4a0e-43c8-96f2-6e749be13fd4