bug 9298: WSS BinarySecurityToken elements with a ValueType URI ending in "#X509PKIPathv1" or "#PKCS7" are now supported by the WSS processor for inbound WSS.
There is still no way provided to generate such tokens for outbound WSS using the WSS decorator (though they could always be generated by hand).


Signatures containing a KeyInfo with a KeyIdentifier with a ValueType of #X509v3 (that is, the key identifier is the actual base-64 encoded certificate itself) are now supported by the WSS processor for inbound WSS, with or without an enclosing SecurityTokenReference (between the KeyInfo and the KeyIdentifier). 

There is still no way provided to generate such signatures for outbound WSS using the WSS decorator (though it is possible that they could be generated by hand by modifying the signature after signing).

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@31376 219f1c6e-4a0e-43c8-96f2-6e749be13fd4