Starting work on embedded Derby database, based on Steve's work in the rel6_1_5_derby branch.
SSG Derby support currently requires a hand-maintained ssg_embedded.sql file containing the database.  This is due to a Hibernate issue with cascading delete with @CollectionOfElements that is tracked in their JIRA as "ANN-747".  Hand-maintaining two SSG SQL files is not considered to be an acceptable solution for Fangtooth and will be replaced with something better (one way or another) before the release ships.

The Gateway creates an embedded database in Gateway/node/default/var/db/ on startup, if it isn't already present.

When using Derby a different spring configuration is used that does not use C3P0 (or require a database username/password).

Hibernate mappings for group memberships are altered to prevent generation of SQL that attempts to compare strings with integers: group and user IDs in GroupMembership are now represented in code as longs instead of strings.  

ESM utility code for managing the Derby embedded DB have been promited to  layer7-gateway-server so the Gateway can use it as well.

Hibernate mappings referring to fields named "xml" have been quoted to avoid conflicts with the Derby keyword "xml".

Config wizard option groups can now be marked as "optional".  It is possible to skip the entire MySQL database setup portion of Gateway configuration menu in which case the Gateway will use an embedded Derby database instead (and will run with a reduced feature set).
  
  There are a number of limitations when the Gateway is run with an embedded Derby database:
  
  * No clustering (obviously): It is not possible to join such a Gateway node into a cluster.
  
  * Sjones noted that "The configured SSM administrative account and the cluster hostname are currently ignored when using an embedded database," but I have not confirmed what is meant by this.
  
  * Audit archiving and Service metrics are disabled when using an embedded database.
  
  * Backup and restore does not currently work with Derby.
  
  * The Derby database log is not rotated and cannot be viewed from the ssgconfig menu.
  
  TODO: Address above limitations that are addressable for Fangtooth, as well as: add additional Derby-specfic exceptions to UnsupportedExceptionsThrowsAdvice, check whether audit downloads work with Derby, produce a framework for generating, maintaining, and updating the Derby schema.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@31479 219f1c6e-4a0e-43c8-96f2-6e749be13fd4