bug 12518: The SSM thick client now installs Crypto-J 6.0.0.1 as its highest-preference crypto provider by default.  (The applet is still on its own, and is stuck with whatever providers are configured in the JDK the user is using to run the applet -- if the only DH provider is limited to 1024 bit primes, then the applet will remain unable to connect using DHE cipher suites to a TLS server that uses larger primes for ephemeral Diffie-Hellman.)
The SSM system property "com.l7tech.console.security.cryptoj.install" can be set to adjust the behavior of the SSM thick-client.  Set it to "first" to install as most-preferred provider; "last" or "true" to install as least-preference provider (as it was pre-Fangtooth); or "false" to avoid installing Crypto-J as a security provider.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@33143 219f1c6e-4a0e-43c8-96f2-6e749be13fd4