bug 13197: Implemented support for custom security interceptors in the @Secured annotation.  Wrote JavaDoc for the various methods in the @Secured annotation.
Private key admin methods now use a custom interceptor supports a new @PrivateKeySecured annotation that can express security policy for working with private keys and their keystores.  Private key admin methods now require appropriate access to both the key entries being worked with and their owning keystores.  The SSG_KEYSTORE entity is now mapped to the SsgKeyFinder entity interface instead of the KeystoreFile entity class (which is not really exposed in the Gateway server code outside of the SsgKeyStoreManagerImpl and the various keystore implementations).  It is now possible for a user to be given permission to create a private key but not to read any.  Other granular access control policies regarding keys and keystores will also now function as intended.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@33478 219f1c6e-4a0e-43c8-96f2-6e749be13fd4