bug 12129: Prototype half-sync/half-async implementation of the ThroughputQuota assertion as a modular assertion (no core changes).  This version of the assertion attempts to provide behavior as close as possible to the original, but uses a different approach for counter updates.  Counter reads are performed immediately, synchronously.  The value is then preincremented and used to determine assertion success/failure.  The actual modification of the counter is done in a background thread that batches up pending modifications into a single commit.  At most one thread at a time (per node) will be responsible for collecting and committing batch updates for a counter.  Decrements are also processed asynchronously (but in arrival order).
Because updates can lag behind, it is possible for the assertion to succeed even though the increment it succeeds on is later rejected for exceeding the limit.  It looks like overages are still detected fairly quickly however.

Overall performance appears to be substantially improved, and errors due to multiple threads using row locking to attempt to modify the same row at the same time can no longer be reproduced by me (and should not be possible on a single Gateway with any amount of concurrency).

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@33638 219f1c6e-4a0e-43c8-96f2-6e749be13fd4