Encapsulated assertion progress.   Encapsulated assertions now show up on the palette and can be used in policies.  The palette updates when an encapsulated assertion is created or edited on the same SSM.  (If one is created or edited on a different SSM, you must log out and back in to see it in the palette.)
Created a centralized registry of palette folders within the SSM, and removed several more special-purpose palette folder nodes.  This is currently not data-driven because of the need to support some legacy palette folder node behavior (in AccessControlFolderNode).  This looks like it is fixable, though.

Selecting an existing icon for an encapsulated assertion now works as expected, though it currently allows you to select icons larger than 16x16.   (Support for specifying your own completely custom image to use as an icon is about 15% complete.)



git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@34105 219f1c6e-4a0e-43c8-96f2-6e749be13fd4