SSG-6325 HostnameVerifyingSSLSocketWrapper implements verify after SSL Handshake.
In this case we permit the JDK LDAPS client kick off the handshake instead of our code forcing it. We then intercept the startHandshake method and verify the hostname after it completes.

This is all so that multiple SSL handshakes are avoided so that WindowsAD servers do not disconnect the LDAP session.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@35200 219f1c6e-4a0e-43c8-96f2-6e749be13fd4