Custom Assertion Enhancements: add ability to license individual Custom Assertions.
The current implementation of Custom Assertion limits licensing to all or none (nothing more granular). Added to this release is the ability to license individual Custom Assertions and maintain backwards compatibility with previous license feature set for Custom Assertions. This is done by supporting an optional property, which can be configured to (fromCustomAssertionClass), this will allow the Custom Assertion class name to license individually from a feature set. If the property is not specified, the Custom Assertion is license enabled (or disabled) with all other Custom Assertions as per previous behavior for backwards compatibility.

In custom_assertions.properties, support property <MyAssertionName>.feature.set.name=(fromCustomAssertionClass).
   - This property is optional, if not specified, the Custom Assertion is license enabled (or disabled) with all other Custom Assertions for the CustomAssertionHolder.class feature. License feature sets are defined in GatewayFeatureSets, see ass(CustomAssertionHolder.class) to enable all Custom Assertions. Developer note: CustomerAssertionHolder meta defaults FEATURE_SET_NAME to (fromClass), which is CustomAssertionHolder.class. This default behaviour must be maintained for backwards compatibility.   
   - If the property is set to (fromCustomAssertionClass), the individual Custom Assertion is license enabled as its own feature. The feature can be added to GatewayFeatureSets (e.g. cass("assertion:SalesForceConnectorCustom")).
   
   https://wiki.l7tech.com/mediawiki/index.php/Custom_Assertion_SDK_Enhancement#Other

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@37146 219f1c6e-4a0e-43c8-96f2-6e749be13fd4