SSG-8805: Encapsulated assertion call sites (ie, EncapsulatedAssertion policy assertion bean XML) can now declare a property noOpIfConfigMissing=true.  If this is set to true, then it is expected that sometimes the encapsulated assertion configuration (identified by GUID) will not be present on the current system, and this is not an error.  Instead the assertion will show as greyed out at design time, and at runtime will always succeed (returning AssertionStatus.NONE) while taking no action.
No way is currently provided to set this property to true via the GUI.  It must be set to true by adding <L7p:NoOpIfConfigMissing booleanValue="true"/> to the <L7p:Encapsulated> element in the policy XML at each call site that should be marked as optional.  

This feature is intended to be used to support shipping a policy that can make use of optional additional  features that are installed with a separate add-on pack.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@40597 219f1c6e-4a0e-43c8-96f2-6e749be13fd4